backend/requirements.txt
flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0
flask-limiter==3.8.0
pydantic==2.9.2
requests==2.32.3
redis==5.0.8
python-dotenv==1.0.1
structlog==24.1.0
google-generativeai==0.7.2
wikipedia-api==0.6.0